feat: move 4 simple FeedFilters to commons (batch 6)
Move HiddenAccountsFeedFilter, HiddenWordsFeedFilter,
FollowPackMembersFeedFilter, and PinnedNotesFeedFilter
from amethyst to commons module.

IAccount additions:
- liveHiddenUsers: LiveHiddenUsers (current hidden users state)
- isKnown(User/String): Boolean (follow status check)
- pinnedNotesList(): List<Note> (pinned notes)

Original files replaced with typealiases for backwards
compatibility. ViewModels updated to pass ICacheProvider
where needed.

Filters not moved (need cache collection access):
- RelayFeedFilter (needs notes.filterIntoSet)
- UserProfileMutualFeedFilter (needs notes + addressables)
- UserProfileAppRecommendationsFeedFilter (needs mapFlattenIntoSet)
- WebBookmarkFeedFilter (needs specialized filterIntoSet extension)
