+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/followPacks/feed/dal/FollowPackMembersFeedFilter.kt
@@ -20,35 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.followPacks.feed.dal
 
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.AddressableNote
-import com.vitorpamplona.amethyst.model.LocalCache.checkGetOrCreateUser
-import com.vitorpamplona.amethyst.model.User
-import com.vitorpamplona.amethyst.ui.dal.FeedFilter
-import com.vitorpamplona.quartz.nip51Lists.followList.FollowListEvent
-
-class FollowPackMembersFeedFilter(
-    val followPackNote: AddressableNote,
-    val account: Account,
-) : FeedFilter<User>() {
-    override fun feedKey(): String = account.userProfile().pubkeyHex + "-" + followPackNote.idHex
-
-    val cache: MutableMap<FollowListEvent, List<User>> = mutableMapOf()
-
-    override fun feed(): List<User> {
-        val followPackEvent = followPackNote.event as? FollowListEvent ?: return emptyList()
-
-        val previousList = cache[followPackEvent]
-        if (previousList != null) return previousList
-
-        val follows =
-            followPackEvent
-                .followIdSet()
-                .mapNotNull { hex -> checkGetOrCreateUser(hex) }
-                .filter { !account.isHidden(it) }
-                .sortedByDescending { account.isKnown(it) }
-
-        cache[followPackEvent] = follows
-        return follows
-    }
-}
+// Re-export from commons for backwards compatibility
+typealias FollowPackMembersFeedFilter = com.vitorpamplona.amethyst.commons.ui.feeds.dal.FollowPackMembersFeedFilter
