+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/settings/dal/HiddenWordsFeedFilter.kt
@@ -20,17 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.settings.dal
 
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.ui.dal.FeedFilter
-
-class HiddenWordsFeedFilter(
-    val account: Account,
-) : FeedFilter<String>() {
-    override fun feedKey(): String = account.userProfile().pubkeyHex
-
-    override fun showHiddenKey(): Boolean = true
-
-    override fun feed(): List<String> =
-        account.hiddenUsers.flow.value.hiddenWords
-            .toList()
-}
+// Re-export from commons for backwards compatibility
+typealias HiddenWordsFeedFilter = com.vitorpamplona.amethyst.commons.ui.feeds.dal.HiddenWordsFeedFilter
