+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/ui/feeds/dal/FollowPackMembersFeedFilter.kt
@@ -1,0 +1,55 @@
+/*
+ * Copyright (c) 2025 Vitor Pamplona
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+ * Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
+ * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package com.vitorpamplona.amethyst.commons.ui.feeds.dal
+
+import com.vitorpamplona.amethyst.commons.model.AddressableNote
+import com.vitorpamplona.amethyst.commons.model.IAccount
+import com.vitorpamplona.amethyst.commons.model.User
+import com.vitorpamplona.amethyst.commons.model.cache.ICacheProvider
+import com.vitorpamplona.amethyst.commons.ui.feeds.FeedFilter
+import com.vitorpamplona.quartz.nip51Lists.followList.FollowListEvent
+
+class FollowPackMembersFeedFilter(
+    val followPackNote: AddressableNote,
+    val account: IAccount,
+    val cache: ICacheProvider,
+) : FeedFilter<User>() {
+    override fun feedKey(): String = account.userProfile().pubkeyHex + "-" + followPackNote.idHex
+
+    val localCache: MutableMap<FollowListEvent, List<User>> = mutableMapOf()
+
+    override fun feed(): List<User> {
+        val followPackEvent = followPackNote.event as? FollowListEvent ?: return emptyList()
+
+        val previousList = localCache[followPackEvent]
+        if (previousList != null) return previousList
+
+        val follows =
+            followPackEvent
+                .followIdSet()
+                .mapNotNull { hex -> cache.getOrCreateUser(hex) }
+                .filter { !account.isHidden(it) }
+                .sortedByDescending { account.isKnown(it) }
+
+        localCache[followPackEvent] = follows
+        return follows
+    }
+}
