+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/ui/feeds/dal/HiddenAccountsFeedFilter.kt
@@ -18,23 +18,31 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.ui.screen.loggedIn.followPacks.feed.dal
+package com.vitorpamplona.amethyst.commons.ui.feeds.dal
 
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.AddressableNote
-import com.vitorpamplona.amethyst.ui.screen.UserFeedViewModel
+import com.vitorpamplona.amethyst.commons.model.IAccount
+import com.vitorpamplona.amethyst.commons.model.User
+import com.vitorpamplona.amethyst.commons.model.cache.ICacheProvider
+import com.vitorpamplona.amethyst.commons.ui.feeds.FeedFilter
+import com.vitorpamplona.quartz.utils.Log
+import kotlinx.coroutines.CancellationException
 
-class FollowPackMembersUserFeedViewModel(
-    val followPackNote: AddressableNote,
-    val account: Account,
-) : UserFeedViewModel(FollowPackMembersFeedFilter(followPackNote, account, account.cache)) {
-    class Factory(
-        val followPackNote: AddressableNote,
-        val account: Account,
-    ) : ViewModelProvider.Factory {
-        @Suppress("UNCHECKED_CAST")
-        override fun <T : ViewModel> create(modelClass: Class<T>): T = FollowPackMembersUserFeedViewModel(followPackNote, account) as T
-    }
+class HiddenAccountsFeedFilter(
+    val account: IAccount,
+    val cache: ICacheProvider,
+) : FeedFilter<User>() {
+    override fun feedKey(): String = account.userProfile().pubkeyHex
+
+    override fun showHiddenKey(): Boolean = true
+
+    override fun feed(): List<User> =
+        account.liveHiddenUsers.hiddenUsers.reversed().mapNotNull {
+            try {
+                cache.getOrCreateUser(it)
+            } catch (e: Exception) {
+                if (e is CancellationException) throw e
+                Log.e("HiddenAccountsFeedFilter") { "Failed to parse key $it" }
+                null
+            }
+        }
 }
