+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/ui/feeds/dal/HiddenWordsFeedFilter.kt
@@ -18,23 +18,19 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.ui.screen.loggedIn.followPacks.feed.dal
+package com.vitorpamplona.amethyst.commons.ui.feeds.dal
 
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.AddressableNote
-import com.vitorpamplona.amethyst.ui.screen.UserFeedViewModel
+import com.vitorpamplona.amethyst.commons.model.IAccount
+import com.vitorpamplona.amethyst.commons.ui.feeds.FeedFilter
 
-class FollowPackMembersUserFeedViewModel(
-    val followPackNote: AddressableNote,
-    val account: Account,
-) : UserFeedViewModel(FollowPackMembersFeedFilter(followPackNote, account, account.cache)) {
-    class Factory(
-        val followPackNote: AddressableNote,
-        val account: Account,
-    ) : ViewModelProvider.Factory {
-        @Suppress("UNCHECKED_CAST")
-        override fun <T : ViewModel> create(modelClass: Class<T>): T = FollowPackMembersUserFeedViewModel(followPackNote, account) as T
-    }
+class HiddenWordsFeedFilter(
+    val account: IAccount,
+) : FeedFilter<String>() {
+    override fun feedKey(): String = account.userProfile().pubkeyHex
+
+    override fun showHiddenKey(): Boolean = true
+
+    override fun feed(): List<String> =
+        account.liveHiddenUsers.hiddenWords
+            .toList()
 }
